Add files via upload
In this economy, loans have become a major requirement in everyone's life. It is beneficial for people who are buying their dream house and for students to keep up with their studies and living expenses. In that case, banks have to look after different aspects of the candidate's profile to see if that applicant is eligible to be granted with a loan or not.
Here I'm using Machine Learning with Python to predict whether the applicant's profile is eligible for the loan or not using some key features.